refactor: use ESM instead of Babel
