--- .babelrc
@@ -1,13 +1,0 @@
-{
-  "presets": [
-    "latest"
-  ],
-  "env": {
-    "development": {
-      "presets": [
-        "jsdoc-to-assert",
-        "power-assert"
-      ]
-    }
-  }
-}
