--- .travis.yml
@@ -1,3 +1,0 @@
-sudo: false
-language: node_js
-node_js: "stable"
