+++ bin/cmd.js
@@ -1,6 +1,8 @@
 #!/usr/bin/env node
+import meow from "meow";
+import concat from "concat-stream";
+import create from "../src/create-textlint-rule-example.js";
 
-const meow = require('meow');
 const cli = meow(`
     Usage
       $ create-textlint-rule-example <file-path>
@@ -8,14 +10,12 @@ const cli = meow(`
     Options
       --separator separator between each examples
 `);
-const create = require("../lib/create-textlint-rule-example");
-const concat = require('concat-stream');
 const fs = require('fs');
 const file = process.argv[2];
 const input = file && file !== '-'
     ? fs.createReadStream(process.argv[2])
     : process.stdin;
-input.pipe(concat(function(buf) {
+input.pipe(concat(function (buf) {
     console.log(create(buf.toString('utf8'), {
         exampleSeparator: cli.flags.separator
     }));
