+++ src/index.js
@@ -1,6 +1,7 @@
 // MIT © 2017 azu
 "use strict";
-const createRuleExample = require("./create-textlint-rule-example");
-const getRuleTest = require("./get-rule-test");
-module.exports.createRuleExample = createRuleExample;
-module.exports.getRuleTest = getRuleTest;
+import createRuleExample from "./create-textlint-rule-example.js";
+import getRuleTest from "./get-rule-test.js";
+
+export { createRuleExample }
+export { getRuleTest }
