+++ src/mock-textlint-tester.js
@@ -1,12 +1,13 @@
 // MIT © 2017 azu
 "use strict";
-const clone = require('clone');
+import clone from "clone";
+
 const calls = {};
-module.exports = class MockTextlintRuleTester {
-    run(ruleName, rule, { valid, invalid }) {
+export default class MockTextlintRuleTester {
+    run(ruleName, rule, {valid, invalid}) {
         calls.ruleName = ruleName;
         calls.valid = clone(valid);
         calls.invalid = clone(invalid);
     }
-};
-module.exports.calls = calls;
+}
+export {calls}
