--- test/fixtures/no-todo/test/mocha.opts
@@ -1,1 +1,0 @@
---compilers js:espower-babel/guess
